fix(kb): port recall dedup to Cursor and Codex
The dedup shipped in the previous commit was Claude-only — the sibling recall
hooks shared recallQuery but not the seen-set, so Cursor and Codex kept
re-showing notes the session already had.

Extracts the state handling into hooks/recall-seen.mjs (readSeen / writeSeen /
idsInPage / seenArgs) and wires all three hosts to it, so the behaviour cannot
drift between them again. Each host passes its own transcript path; a host that
exposes none (ephemeral Codex runs) simply never resets, which is the safe
direction — dedup still holds for the session.

Also drops a redundant inner `const sid` in codex-kb-recall.mjs now that the
sanitized id is bound once at the top of the handler.

Verified E2E on all three hooks: inject -> silent -> transcript shrinks ->
inject. Codex without a transcript path: inject -> silent. 626 tests pass.
